+++ server/src/test/java/org/apache/druid/server/compaction/CompactionCandidateTest.java
@@ -1,0 +1,141 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.druid.server.compaction;
+
+import org.apache.druid.error.DruidException;
+import org.apache.druid.java.util.common.granularity.Granularities;
+import org.apache.druid.server.coordinator.CreateDataSegments;
+import org.apache.druid.timeline.DataSegment;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.util.Collections;
+import java.util.List;
+
+public class CompactionCandidateTest
+{
+  private static final String DATASOURCE = "test_datasource";
+
+  @Test
+  public void testConstructorAndGetters()
+  {
+    List<DataSegment> segments = createTestSegments(3);
+    CompactionCandidate.ProposedCompaction proposed = CompactionCandidate.ProposedCompaction.from(segments, null);
+    CompactionEligibility eligibility = CompactionEligibility.fail("test reason");
+    CompactionStatus status = CompactionStatus.COMPLETE;
+
+    CompactionCandidate candidate = new CompactionCandidate(proposed, eligibility, status);
+
+    Assert.assertEquals(proposed, candidate.getProposedCompaction());
+    Assert.assertEquals(eligibility, candidate.getPolicyEligibility());
+    Assert.assertEquals(status, candidate.getCurrentStatus());
+    Assert.assertEquals(segments, candidate.getSegments());
+    Assert.assertEquals(DATASOURCE, candidate.getDataSource());
+    Assert.assertEquals(3, candidate.numSegments());
+  }
+
+  @Test
+  public void testProposedCompactionFrom()
+  {
+    List<DataSegment> segments = createTestSegments(3);
+
+    CompactionCandidate.ProposedCompaction proposed =
+        CompactionCandidate.ProposedCompaction.from(segments, null);
+
+    Assert.assertEquals(segments, proposed.getSegments());
+    Assert.assertEquals(DATASOURCE, proposed.getDataSource());
+    Assert.assertEquals(3, proposed.numSegments());
+    Assert.assertNotNull(proposed.getUmbrellaInterval());
+    Assert.assertNotNull(proposed.getCompactionInterval());
+    Assert.assertNotNull(proposed.getStats());
+  }
+
+  @Test
+  public void testProposedCompactionWithTargetGranularity()
+  {
+    List<DataSegment> segments = createTestSegments(5);
+
+    CompactionCandidate.ProposedCompaction proposed =
+        CompactionCandidate.ProposedCompaction.from(segments, Granularities.MONTH);
+
+    Assert.assertEquals(segments, proposed.getSegments());
+    Assert.assertEquals(5, proposed.numSegments());
+    Assert.assertNotNull(proposed.getUmbrellaInterval());
+    Assert.assertNotNull(proposed.getCompactionInterval());
+  }
+
+  @Test
+  public void testProposedCompactionThrowsOnNullOrEmptySegments()
+  {
+    Assert.assertThrows(
+        DruidException.class,
+        () -> CompactionCandidate.ProposedCompaction.from(null, null)
+    );
+
+    Assert.assertThrows(
+        DruidException.class,
+        () -> CompactionCandidate.ProposedCompaction.from(Collections.emptyList(), null)
+    );
+  }
+
+  @Test
+  public void testWithCurrentStatus()
+  {
+    List<DataSegment> segments = createTestSegments(2);
+    CompactionCandidate.ProposedCompaction proposed = CompactionCandidate.ProposedCompaction.from(segments, null);
+    CompactionEligibility eligibility = CompactionEligibility.fail("test");
+    CompactionStatus originalStatus = CompactionStatus.COMPLETE;
+
+    CompactionCandidate candidate = new CompactionCandidate(proposed, eligibility, originalStatus);
+    CompactionStatus newStatus = CompactionStatus.pending("needs compaction");
+    CompactionCandidate updated = candidate.withCurrentStatus(newStatus);
+
+    Assert.assertNotSame(candidate, updated);
+    Assert.assertEquals(originalStatus, candidate.getCurrentStatus());
+    Assert.assertEquals(newStatus, updated.getCurrentStatus());
+    Assert.assertEquals(proposed, updated.getProposedCompaction());
+    Assert.assertEquals(eligibility, updated.getPolicyEligibility());
+  }
+
+  @Test
+  public void testDelegationMethods()
+  {
+    List<DataSegment> segments = createTestSegments(3);
+    CompactionCandidate.ProposedCompaction proposed = CompactionCandidate.ProposedCompaction.from(segments, null);
+    CompactionCandidate candidate = new CompactionCandidate(
+        proposed,
+        CompactionEligibility.fail("test"),
+        CompactionStatus.COMPLETE
+    );
+
+    Assert.assertEquals(proposed.getTotalBytes(), candidate.getTotalBytes());
+    Assert.assertEquals(proposed.getUmbrellaInterval(), candidate.getUmbrellaInterval());
+    Assert.assertEquals(proposed.getCompactionInterval(), candidate.getCompactionInterval());
+    Assert.assertEquals(proposed.getStats().getTotalBytes(), candidate.getStats().getTotalBytes());
+  }
+
+  private static List<DataSegment> createTestSegments(int count)
+  {
+    return CreateDataSegments.ofDatasource(DATASOURCE)
+                             .forIntervals(count, Granularities.DAY)
+                             .startingAt("2024-01-01")
+                             .eachOfSizeInMb(100);
+  }
+}
