+++ server/src/test/java/org/apache/druid/server/coordinator/duty/CompactSegmentsTest.java
@@ -79,7 +79,7 @@ import org.apache.druid.rpc.indexing.OverlordClient;
 import org.apache.druid.segment.IndexSpec;
 import org.apache.druid.segment.incremental.OnheapIncrementalIndex;
 import org.apache.druid.segment.transform.CompactionTransformSpec;
-import org.apache.druid.server.compaction.CompactionCandidate;
+import org.apache.druid.server.compaction.CompactionCandidate.ProposedCompaction;
 import org.apache.druid.server.compaction.CompactionCandidateSearchPolicy;
 import org.apache.druid.server.compaction.CompactionSlotManager;
 import org.apache.druid.server.compaction.CompactionStatusTracker;
@@ -876,7 +876,7 @@ public class CompactSegmentsTest
     // All segments is compact at the same time since we changed the segment granularity to YEAR and all segment
     // are within the same year
     Assert.assertEquals(
-        CompactionCandidate.from(datasourceToSegments.get(dataSource), Granularities.YEAR).getCompactionInterval(),
+        ProposedCompaction.from(datasourceToSegments.get(dataSource), Granularities.YEAR).getCompactionInterval(),
         taskPayload.getIoConfig().getInputSpec().getInterval()
     );
 
@@ -1071,7 +1071,7 @@ public class CompactSegmentsTest
     // All segments is compact at the same time since we changed the segment granularity to YEAR and all segment
     // are within the same year
     Assert.assertEquals(
-        CompactionCandidate.from(datasourceToSegments.get(dataSource), Granularities.YEAR).getCompactionInterval(),
+        ProposedCompaction.from(datasourceToSegments.get(dataSource), Granularities.YEAR).getCompactionInterval(),
         taskPayload.getIoConfig().getInputSpec().getInterval()
     );
 
@@ -1171,7 +1171,7 @@ public class CompactSegmentsTest
     // All segments is compact at the same time since we changed the segment granularity to YEAR and all segment
     // are within the same year
     Assert.assertEquals(
-        CompactionCandidate.from(datasourceToSegments.get(dataSource), Granularities.YEAR).getCompactionInterval(),
+        ProposedCompaction.from(datasourceToSegments.get(dataSource), Granularities.YEAR).getCompactionInterval(),
         taskPayload.getIoConfig().getInputSpec().getInterval()
     );
 
@@ -1408,7 +1408,7 @@ public class CompactSegmentsTest
     ClientCompactionTaskQuery taskPayload = (ClientCompactionTaskQuery) payloadCaptor.getValue();
 
     Assert.assertEquals(
-        CompactionCandidate.from(segments, Granularities.DAY).getCompactionInterval(),
+        ProposedCompaction.from(segments, Granularities.DAY).getCompactionInterval(),
         taskPayload.getIoConfig().getInputSpec().getInterval()
     );
 
@@ -1471,7 +1471,7 @@ public class CompactSegmentsTest
     ClientCompactionTaskQuery taskPayload = (ClientCompactionTaskQuery) payloadCaptor.getValue();
 
     Assert.assertEquals(
-        CompactionCandidate.from(segments, Granularities.YEAR).getCompactionInterval(),
+        ProposedCompaction.from(segments, Granularities.YEAR).getCompactionInterval(),
         taskPayload.getIoConfig().getInputSpec().getInterval()
     );
